+++ ichnaea/async/schedule.py
@@ -13,6 +13,11 @@ CELERYBEAT_SCHEDULE = {
         'schedule': timedelta(seconds=60),
         'options': {'expires': 57},
     },
+    'monitor-api-key-limits': {
+        'task': 'ichnaea.monitor.tasks.monitor_api_key_limits',
+        'schedule': timedelta(seconds=600),
+        'options': {'expires': 570},
+    },
 
     # Continuous location update tasks